Moving is_regularly_spaced function to post_processing.py
